added access level to users and super users, making sure users do not have access to about information about other users under a different super user and even within under the same super user, each low level user enjoys privacy to their information

+++ server/common/controllers/super.user.controller.ts
@@ -2,17 +2,23 @@ import { Request, Response } from 'express';
 import { superUserCrud } from '../models/super.user.js';
 
 export const superUserController = {
-  getUser: (req: Request, res: Response) => {
+  getUser: (req: any, res: Response) => {
+    const reqId = req.user.userId;
     const {
       params: { userId }
     } = req;
+
+    if (reqId !== userId) {
+      return res.status(500).json({
+        status: false,
+        error: `This endpoint is above the pay grad of user ${reqId}`
+      })
+    }
     superUserCrud.findUser({ id: userId })
       .then((user) => {
-        console.log('superUser>>>>', user?.getUsers().then((e) => console.log('super', e)))
         return res.status(200).json({
           status: true,
           data: user?.toJSON(),
-          cy: user?.countUsers(),
         });
       })
       .catch((err) => {
@@ -22,11 +28,195 @@ export const superUserController = {
         });
       });
   },
-  updateUser: (req: Request, res: Response) => {
+
+  getUsers: (req: Request, res: Response) => {
+    const {
+      params: { userId }
+    } = req;
+
+    superUserCrud.findUser({ id: userId })
+      .then((user) => {
+        user?.getUsers()
+          .then((users) => {
+            return res.status(200).json({
+              status: true,
+              users: users,
+            });
+          })
+          .catch((err) => {
+            return res.status(500).json({
+              status: false,
+              error: err,
+            });
+          });
+      })
+      .catch((err) => {
+        return res.status(500).json({
+          status: false,
+          error: err,
+        });
+      });
+  },
+
+  getUsersCount: (req: Request, res: Response) => {
+    const {
+      params: { userId }
+    } = req;
+
+    superUserCrud.findUser({ id: userId })
+      .then((user) => {
+        user?.countUsers()
+          .then((users) => {
+            return res.status(200).json({
+              status: true,
+              totalNumberOfUsers: users,
+            })
+          })
+          .catch((err) => {
+            return res.status(500).json({
+              status: false,
+              error: err,
+            });
+          });
+      })
+      .catch((err) => {
+        return res.status(500).json({
+          status: false,
+          error: err,
+        });
+      });
+  },
+
+  getDepartments: (req: Request, res: Response) => {
+    const {
+      params: { userId }
+    } = req;
+
+    superUserCrud.findUser({ id: userId })
+      .then((user) => {
+        user?.getDepartments()
+          .then((dept) => {
+            return res.status(200).json({
+              status: true,
+              dept: dept,
+            })
+          })
+          .catch((err) => {
+            return res.status(500).json({
+              status: false,
+              error: err,
+            });
+          });
+      })
+      .catch((err) => {
+        return res.status(500).json({
+          status: false,
+          error: err,
+        });
+      });
+  },
+
+  getDepartMentsCount: (req: Request, res: Response) => {
+    const {
+      params: { userId }
+    } = req;
+
+    superUserCrud.findUser({ id: userId })
+      .then((user) => {
+        user?.countDepartments()
+          .then((departments) => {
+            return res.status(200).json({
+              status: true,
+              totalNumberOfDepartments: departments,
+            });
+          })
+          .catch((err) => {
+            return res.status(500).json({
+              status: false,
+              error: err,
+            });
+          });
+      })
+      .catch((err) => {
+        return res.status(500).json({
+          status: false,
+          error: err,
+        })
+      })
+  },
+
+  getTasks: (req: Request, res: Response) => {
+    const {
+      params: { userId }
+    } = req;
+
+    superUserCrud.findUser({ id: userId })
+      .then((user) => {
+        user?.getTasks()
+          .then((tasks) => {
+            return res.status(200).json({
+              status: true,
+              tasks: tasks,
+            })
+          })
+          .catch((err) => {
+            return res.status(500).json({
+              status: false,
+              error: err,
+            });
+          });
+      })
+      .catch((err) => {
+        return res.status(500).json({
+          status: false,
+          error: err,
+        });
+      });
+  },
+
+  getTasksCount: (req: Request, res: Response) => {
+    const {
+      params: { userId }
+    } = req;
+
+    superUserCrud.findUser({ id: userId })
+      .then((user) => {
+        user?.countTasks()
+          .then((tasks) => {
+            return res.status(200).json({
+              status: true,
+              totalNumberOfTasks: tasks,
+            })
+          })
+          .catch((err) => {
+            return res.status(200).json({
+              status: true,
+              error: err,
+            });
+          });
+      })
+      .catch((err) => {
+        return res.status(500).json({
+          status: false,
+          error: err,
+        });
+      });
+  },
+
+  updateUser: (req: any, res: Response) => {
+    const reqId = req.user.userId;
     const {
       params: { userId },
       body: payload,
     } = req;
+
+    if (reqId !== userId) {
+      return res.status(500).json({
+        status: false,
+        error: `This endpoint is above the pay grad of user ${reqId}`
+      })
+    }
+
     // If the payload doesn't have any keys, return error
     if (!Object.keys(payload).length) {
       return res.status(400).json({
@@ -53,10 +243,19 @@ export const superUserController = {
         });
       });
   },
-  deleteUser: (req: Request, res: Response) => {
+  deleteUser: (req: any, res: Response) => {
+    const reqId = req.user.userId;
     const {
       params: { id: userId }
     } = req;
+
+    if (reqId !== userId) {
+      return res.status(500).json({
+        status: false,
+        error: `This endpoint is above the pay grad of user ${reqId}`
+      })
+    }
+    
     superUserCrud.deleteUser({ id: userId })
       .then((numberOfUsersDeleted) => {
         return res.status(200).json({
